--warning, aaaah, now isn't that better? ;-)
svn path=/trunk/kdeextragear-1/amarok/; revision=384963

+++ src/playlistbrowser.cpp
@@ -1155,7 +1155,7 @@ void PlaylistBrowserItem::removeTrack( QListViewItem *item )
 
 void PlaylistBrowserItem::customEvent( QCustomEvent *e )
 {
-    if( e->type() == PlaylistReader::JobFinishedEvent )
+    if( e->type() == (int)PlaylistReader::JobFinishedEvent )
     {
         foreachType( BundleList, static_cast<PlaylistReader*>(e)->bundles ) {
            const MetaBundle &b = *it;
